Added queue clear.

+++ commands/music/queue.js
@@ -18,29 +18,38 @@ module.exports = class QueueCommand extends commando.Command {
     const queue = await helper.retrieveQueue(server.id);
     queue.songs.shift();
 
-    if (command) {
-      switch (command) {
-        case "edit":
-          helper.createPagination(queue.songs, message, false, true);
-          break;
+    if (queue.songs.length) {
+      if (command) {
+        switch (command) {
+          case "edit":
+            helper.createPagination(queue.songs, message, false, true);
+            break;
 
-        case "clear":
-          const queueLengthBeforeClear = queue.songs.length;
-          queue.songs.forEach(song =>
-            models.SongQueue.destroy({
-              where: { queueId: queue.id, songId: song.id }
-            })
-          );
-          message.channel.send(
-            `Successfully removed ${queueLengthBeforeClear} songs from Queue.`
-          ).then(message => message.delete(5000))
-          break;
+          case "clear":
+            if (!queue.songs.length) {
+              return;
+            }
+            const queueLengthBeforeClear = queue.songs.length;
+            queue.songs.forEach(song =>
+              models.SongQueue.destroy({
+                where: { queueId: queue.id, songId: song.id }
+              })
+            );
+            message.channel
+              .send(
+                `Successfully removed ${queueLengthBeforeClear} songs from Queue.`
+              )
+              .then(message => message.delete(5000));
+            break;
 
-        default:
-          break;
+          default:
+            break;
+        }
+      } else {
+        helper.createPagination(queue.songs, message);
       }
     } else {
-      helper.createPagination(queue.songs, message);
+      message.channel.send("No songs in queue.").then(message => message.delete(5000));
     }
   }
 };
